requirements.txt
Flask>=3.0
pillow
Flask-Mail
pytesseract
requests
ipython
